move breadbox tests
https://coveord.atlassian.net/browse/KIT-1744
