+++ packages/atomic/cypress/integration/facets/timeframe-facet/timeframe-facet.cypress.ts
@@ -22,13 +22,13 @@ import {
 import * as CommonAssertions from '../../common-assertions';
 import * as CommonFacetAssertions from '../facet-common-assertions';
 import * as TimeframeFacetAssertions from './timeframe-facet-assertions';
-import * as BreadboxAssertions from '../../breadbox/breadbox-assertions';
-import {breadboxComponent} from '../../breadbox/breadbox-selectors';
+import * as BreadboxAssertions from '../../breadbox-assertions';
+import {breadboxComponent} from '../../breadbox-selectors';
 import {
   addBreadbox,
   breadboxLabel,
   deselectBreadcrumbAtIndex,
-} from '../../breadbox/breadbox-actions';
+} from '../../breadbox-actions';
 import {addFacet} from '../facet/facet-actions';
 import {FacetSelectors} from '../facet/facet-selectors';
